% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-package.R
\docType{package}
\name{normalize-package}
\alias{normalize-package}
\alias{_PACKAGE}
\title{normalize: Centering and Scaling of Numeric Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a simple method for centering and scaling of numeric data. Certain columns or rows can be ignored when normalizing or be normalized jointly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/loelschlaeger/normalize}
  \item Report bugs at \url{https://github.com/loelschlaeger/normalize/issues}
}

}
\author{
\strong{Maintainer}: Lennart Oelschläger \email{oelschlaeger.lennart@gmail.com}

}
\keyword{internal}
