% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-package.R
\docType{package}
\name{normalize-package}
\alias{normalize-package}
\alias{_PACKAGE}
\title{normalize: Centering and Scaling of Numeric Data}
\description{
Centering and scaling of numeric data with some extensions to the base R solution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/loelschlaeger/normalize}
  \item Report bugs at \url{https://github.com/loelschlaeger/normalize/issues}
}

}
\keyword{internal}
